refactored controller to remove code duplicates
