+++ src/Controller/TimesheetController.php
@@ -23,15 +23,11 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * Controller used to manage timesheets.
- *
  * @Route(path="/timesheet")
  * @Security("is_granted('view_own_timesheet')")
  */
-class TimesheetController extends AbstractController
+class TimesheetController extends TimesheetAbstractController
 {
-    use TimesheetControllerTrait;
-
     /**
      * @Route(path="/", defaults={"page": 1}, name="timesheet", methods={"GET"})
      * @Route(path="/page/{page}", requirements={"page": "[1-9]\d*"}, name="timesheet_paginated", methods={"GET"})
@@ -43,43 +39,7 @@ class TimesheetController extends AbstractController
      */
     public function indexAction($page, Request $request)
     {
-        $query = new TimesheetQuery();
-        $query->setPage($page);
-
-        $form = $this->getToolbarForm($query);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            /** @var TimesheetQuery $query */
-            $query = $form->getData();
-            if (null !== $query->getBegin()) {
-                $query->getBegin()->setTime(0, 0, 0);
-            }
-            if (null !== $query->getEnd()) {
-                $query->getEnd()->setTime(23, 59, 59);
-            }
-        }
-
-        $query->setUser($this->getUser());
-
-        if ($query->hasTags()) {
-            $query->setTags(
-                new ArrayCollection(
-                    $this->getDoctrine()->getRepository(Tag::class)->findIdsByTagNameList(implode(',', $query->getTags()->toArray()))
-                )
-            );
-        }
-
-        /* @var $entries Pagerfanta */
-        $entries = $this->getRepository()->findByQuery($query);
-
-        return $this->render('timesheet/index.html.twig', [
-            'entries' => $entries,
-            'page' => $query->getPage(),
-            'query' => $query,
-            'showFilter' => $form->isSubmitted(),
-            'toolbarForm' => $form->createView(),
-            'showSummary' => $this->getUser()->getPreferenceValue('timesheet.daily_stats', false),
-        ]);
+        return $this->index($page, $request, 'timesheet/index.html.twig');
     }
 
     /**
@@ -91,37 +51,34 @@ class TimesheetController extends AbstractController
      */
     public function exportAction(Request $request)
     {
-        $query = new TimesheetQuery();
-        $query->setResultType(TimesheetQuery::RESULT_TYPE_OBJECTS);
+        return $this->export($request, 'timesheet/export.html.twig');
+    }
 
-        $form = $this->getToolbarForm($query);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            /** @var TimesheetQuery $query */
-            $query = $form->getData();
-        }
+    /**
+     * @Route(path="/{id}/edit", name="timesheet_edit", methods={"GET", "POST"})
+     * @Security("is_granted('edit', entry)")
+     *
+     * @param Timesheet $entry
+     * @param Request $request
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
+     */
+    public function editAction(Timesheet $entry, Request $request)
+    {
+        return $this->edit($entry, $request, 'timesheet/edit.html.twig');
+    }
 
-        // by default the current month is exported, but it can be overwritten
-        if (null === $query->getBegin()) {
-            $query->setBegin($this->dateTime->createDateTime('first day of this month'));
-        }
-        $query->getBegin()->setTime(0, 0, 0);
-
-        if (null === $query->getEnd()) {
-            $query->setEnd($this->dateTime->createDateTime('last day of this month'));
-        }
-        $query->getEnd()->setTime(23, 59, 59);
-
-        // user timesheet always export for the session user
-        $query->setUser($this->getUser());
-
-        /* @var $entries Pagerfanta */
-        $entries = $this->getRepository()->findByQuery($query);
-
-        return $this->render('timesheet/export.html.twig', [
-            'entries' => $entries,
-            'query' => $query,
-        ]);
+    /**
+     * @Route(path="/create", name="timesheet_create", methods={"GET", "POST"})
+     * @Security("is_granted('create_own_timesheet')")
+     *
+     * @param Request $request
+     * @param ProjectRepository $projectRepository
+     * @param ActivityRepository $activityRepository
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
+     */
+    public function createAction(Request $request, ProjectRepository $projectRepository, ActivityRepository $activityRepository)
+    {
+        return $this->create($request, 'timesheet/edit.html.twig', $projectRepository, $activityRepository);
     }
 
     /**
@@ -141,77 +98,5 @@ class TimesheetController extends AbstractController
                 'soft_limit' => $this->getSoftLimit(),
             ]
         );
-    }
-
-    /**
-     * @Route(path="/{id}/edit", name="timesheet_edit", methods={"GET", "POST"})
-     * @Security("is_granted('edit', entry)")
-     *
-     * @param Timesheet $entry
-     * @param Request $request
-     * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
-     */
-    public function editAction(Timesheet $entry, Request $request)
-    {
-        return $this->edit($entry, $request, 'timesheet', 'timesheet/edit.html.twig');
-    }
-
-    /**
-     * @Route(path="/create", name="timesheet_create", methods={"GET", "POST"})
-     * @Security("is_granted('create_own_timesheet')")
-     *
-     * @param Request $request
-     * @param ProjectRepository $projectRepository
-     * @param ActivityRepository $activityRepository
-     * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
-     */
-    public function createAction(Request $request, ProjectRepository $projectRepository, ActivityRepository $activityRepository)
-    {
-        return $this->create($request, 'timesheet', 'timesheet/edit.html.twig', $projectRepository, $activityRepository);
-    }
-
-    /**
-     * @param Timesheet $entry
-     * @return \Symfony\Component\Form\FormInterface
-     */
-    protected function getCreateForm(Timesheet $entry)
-    {
-        return $this->createForm(TimesheetEditForm::class, $entry, [
-            'action' => $this->generateUrl('timesheet_create', []),
-            'include_rate' => $this->isGranted('edit_rate', $entry),
-            'customer' => true,
-        ]);
-    }
-
-    /**
-     * @param Timesheet $entry
-     * @param int $page
-     * @return \Symfony\Component\Form\FormInterface
-     */
-    protected function getEditForm(Timesheet $entry, $page)
-    {
-        return $this->createForm(TimesheetEditForm::class, $entry, [
-            'action' => $this->generateUrl('timesheet_edit', [
-                'id' => $entry->getId(),
-                'page' => $page,
-            ]),
-            'include_rate' => $this->isGranted('edit_rate', $entry),
-            'include_exported' => $this->isGranted('edit_export', $entry),
-            'customer' => true,
-        ]);
-    }
-
-    /**
-     * @param TimesheetQuery $query
-     * @return \Symfony\Component\Form\FormInterface
-     */
-    protected function getToolbarForm(TimesheetQuery $query)
-    {
-        return $this->createForm(TimesheetToolbarForm::class, $query, [
-            'action' => $this->generateUrl('timesheet', [
-                'page' => $query->getPage(),
-            ]),
-            'method' => 'GET',
-        ]);
     }
 }
